Introduce initial compartments.
Change the makefile to compile every .c file into a .so file
directly. Also experiment and show that functions can be
made private, displaying that by using: `LuaV_execute` and
making it `__attribute__ ((visibility ("hidden")))`. This means
that it wont be exported in the global offset table and will be
truly private because of CHERI and Sentry Capabilities.