Compile the lua interpreter into LLVM bitcode.
This is useful for analyzing the structure of the interpreter
and getting the call graph without going through the c source
files. It is also possible to obtain other information using
the llvm `opt` tool to print various data.

For example: `opt --print-callgraph`.
`opt --help | grep print` to obtain all the printing options.